Renamed to follow convention and added transformation definition entry
