+++ etor/src/main/resources/transformation_definitions.json
@@ -117,6 +117,21 @@
           "args": { }
         }
       ]
+    },
+    {
+      "name": "ucsdOruRemovePatientNameTypeCode",
+      "description": "Removes UCSD ORU's Name Type Code (PID-5.7) from Patient Name (PID-5)",
+      "message": "",
+      "conditions": [
+        "Bundle.entry.resource.ofType(MessageHeader).event.code = 'R01'",
+        "Bundle.entry.resource.ofType(MessageHeader).destination.receiver.resolve().identifier.where(extension.value = 'HD.2,HD.3').value = 'UCSD'"
+      ],
+      "rules": [
+        {
+          "name": "RemovePatientNameTypeCode",
+          "args": { }
+        }
+      ]
     }
   ]
 }
